+++ Classes/UVConfig.m
@@ -18,6 +18,7 @@
 @synthesize email;
 @synthesize displayName;
 @synthesize guid;
+@synthesize customFields;
 
 + (UVConfig *)configWithSite:(NSString *)site andKey:(NSString *)key andSecret:(NSString *)secret {
     return [[[UVConfig alloc] initWithSite:site andKey:key andSecret:secret] autorelease];