+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## master ##
 
+* Add support for setting custom fields programmatically
+
 ## 1.2.3 ##
 
 * Fix a bug that would cause sign-in to fail when connected to a private forum